+++ packages/perseus-editor/src/components/util.ts
@@ -1,7 +1,10 @@
+import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
+
 import type {
     LockedFigure,
     LockedFigureType,
-    LockedPoint,
+    LockedPointType,
+    LockedLineType,
 } from "@khanacademy/perseus";
 
 export function focusWithChromeStickyFocusBugWorkaround(element: Element) {
@@ -44,19 +47,41 @@ export function focusWithChromeStickyFocusBugWorkaround(element: Element) {
 }
 
 export function getValidNumberFromString(value: string) {
-    const parsed = parseInt(value);
+    const parsed = parseFloat(value);
     // If the value is not a number, return 0.
     return isNaN(parsed) ? 0 : parsed;
 }
 
-export function getDefaultFigureForFigureType(
-    type: LockedFigureType,
-): LockedFigure {
+export function getDefaultFigureForType(type: "point"): LockedPointType;
+export function getDefaultFigureForType(type: "line"): LockedLineType;
+export function getDefaultFigureForType(type: LockedFigureType): LockedFigure;
+export function getDefaultFigureForType(type: LockedFigureType): LockedFigure {
     switch (type) {
         case "point":
             return {
                 type: "point",
                 coord: [0, 0],
-            } as LockedPoint;
+                color: "blue",
+                filled: true,
+            };
+        case "line":
+            return {
+                type: "line",
+                kind: "line",
+                points: [
+                    getDefaultFigureForType("point"),
+                    {
+                        ...getDefaultFigureForType("point"),
+                        coord: [2, 2],
+                    },
+                ],
+                color: "blue",
+                lineStyle: "solid",
+                showArrows: false,
+                showStartPoint: false,
+                showEndPoint: false,
+            };
+        default:
+            throw new UnreachableCaseError(type);
     }
 }