+++ packages/perseus/src/widgets/__testdata__/interactive-graph.testdata.ts
@@ -200,7 +200,6 @@ export const linearQuestionWithDefaultCorrect: PerseusRenderer = {
     },
 };
 
-const bluePointStyle = {stroke: "#6494ed", fill: "#6494ed"};
 export const linearQuestionWithLockedPoints: PerseusRenderer = {
     content:
         "**Draw the line of reflection that will map $\\triangle{SIM}$ onto the other triangle below.**\n\n\n[[☃ interactive-graph 1]]",
@@ -226,17 +225,20 @@ export const linearQuestionWithLockedPoints: PerseusRenderer = {
                     {
                         type: "point",
                         coord: [5, 3],
-                        style: bluePointStyle,
+                        color: "blue",
+                        filled: true,
                     },
                     {
                         type: "point",
                         coord: [4, -4],
-                        style: bluePointStyle,
+                        color: "blue",
+                        filled: true,
                     },
                     {
                         type: "point",
                         coord: [7, -3],
-                        style: bluePointStyle,
+                        color: "blue",
+                        filled: true,
                     },
                 ],
                 markings: "graph",
@@ -966,10 +968,14 @@ export const segmentWithLockedPointsQuestion: PerseusRenderer = {
                     {
                         type: "point",
                         coord: [-7, -7],
+                        color: "blue",
+                        filled: true,
                     },
                     {
                         type: "point",
                         coord: [2, -5],
+                        color: "blue",
+                        filled: false,
                     },
                 ],
             },
@@ -1019,12 +1025,105 @@ export const segmentWithLockedPointsWithColorQuestion: PerseusRenderer = {
                     {
                         type: "point",
                         coord: [-7, -7],
-                        style: {fill: "green", stroke: "green"},
+                        color: "green",
+                        filled: true,
                     },
                     {
                         type: "point",
                         coord: [2, -5],
-                        style: {fill: "green", stroke: "green"},
+                        color: "green",
+                        filled: true,
+                    },
+                ],
+            },
+            type: "interactive-graph",
+            version: {
+                major: 0,
+                minor: 0,
+            },
+        },
+    },
+};
+
+export const segmentWithLockedLineQuestion: PerseusRenderer = {
+    content:
+        "Line segment $\\overline{OG}$ is rotated $180^\\circ$ about the point $(-2,4)$.  \n\n**Draw the image of this rotation using the interactive graph.**\n\n*The direction of a rotation by a positive angle is counter-clockwise.* \n\n[[☃ interactive-graph 1]]\n\n",
+    images: {},
+    widgets: {
+        "interactive-graph 1": {
+            graded: true,
+            options: {
+                correct: {
+                    coords: [
+                        [
+                            [-7, -7],
+                            [2, -5],
+                        ],
+                    ],
+                    type: "segment",
+                },
+                graph: {
+                    type: "segment",
+                },
+                gridStep: [1, 1],
+                labels: ["x", "y"],
+                markings: "graph",
+                range: [
+                    [-10, 10],
+                    [-10, 10],
+                ],
+                rulerLabel: "",
+                rulerTicks: 10,
+                showProtractor: false,
+                showRuler: false,
+                snapStep: [0.5, 0.5],
+                step: [1, 1],
+                lockedFigures: [
+                    {
+                        type: "line",
+                        kind: "line",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -7],
+                                color: "purple",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -5],
+                                color: "purple",
+                                filled: false,
+                            },
+                        ],
+                        color: "purple",
+                        lineStyle: "solid",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: true,
+                    },
+                    {
+                        type: "line",
+                        kind: "segment",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -6],
+                                color: "green",
+                                filled: false,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -4],
+                                color: "green",
+                                filled: true,
+                            },
+                        ],
+                        color: "green",
+                        lineStyle: "solid",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: true,
                     },
                 ],
             },
@@ -1307,3 +1406,237 @@ export const questionsAndAnswers: ReadonlyArray<
         ],
     ],
 ];
+
+export const segmentWithAllLockedLineSegmentVariations: PerseusRenderer = {
+    content: "All locked line segments\n\n[[☃ interactive-graph 1]]",
+    images: {},
+    widgets: {
+        "interactive-graph 1": {
+            graded: true,
+            options: {
+                correct: {
+                    coords: [
+                        [
+                            [-7, -7],
+                            [2, -5],
+                        ],
+                    ],
+                    type: "segment",
+                },
+                graph: {
+                    type: "segment",
+                },
+                gridStep: [1, 1],
+                labels: ["x", "y"],
+                markings: "graph",
+                range: [
+                    [-10, 10],
+                    [-10, 10],
+                ],
+                rulerLabel: "",
+                rulerTicks: 10,
+                showProtractor: false,
+                showRuler: false,
+                snapStep: [0.5, 0.5],
+                step: [1, 1],
+                lockedFigures: [
+                    // Point shown, one filled, one open
+                    {
+                        type: "line",
+                        kind: "segment",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -5],
+                                color: "green",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -3],
+                                color: "green",
+                                filled: false,
+                            },
+                        ],
+                        color: "green",
+                        lineStyle: "solid",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: true,
+                    },
+                    // Dashed line, one point shown
+                    {
+                        type: "line",
+                        kind: "segment",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -6],
+                                color: "blue",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -4],
+                                color: "purple",
+                                filled: false,
+                            },
+                        ],
+                        color: "purple",
+                        lineStyle: "dashed",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: false,
+                    },
+                    // Show arrows
+                    {
+                        type: "line",
+                        kind: "segment",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -7],
+                                color: "blue",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -5],
+                                color: "blue",
+                                filled: false,
+                            },
+                        ],
+                        color: "blue",
+                        lineStyle: "solid",
+                        showArrows: true,
+                        showStartPoint: false,
+                        showEndPoint: false,
+                    },
+                ],
+            },
+            type: "interactive-graph",
+            version: {
+                major: 0,
+                minor: 0,
+            },
+        },
+    },
+};
+
+export const segmentWithAllLockedLineVariations: PerseusRenderer = {
+    content: "All locked lines\n\n[[☃ interactive-graph 1]]",
+    images: {},
+    widgets: {
+        "interactive-graph 1": {
+            graded: true,
+            options: {
+                correct: {
+                    coords: [
+                        [
+                            [-7, -7],
+                            [2, -5],
+                        ],
+                    ],
+                    type: "segment",
+                },
+                graph: {
+                    type: "segment",
+                },
+                gridStep: [1, 1],
+                labels: ["x", "y"],
+                markings: "graph",
+                range: [
+                    [-10, 10],
+                    [-10, 10],
+                ],
+                rulerLabel: "",
+                rulerTicks: 10,
+                showProtractor: false,
+                showRuler: false,
+                snapStep: [0.5, 0.5],
+                step: [1, 1],
+                lockedFigures: [
+                    // Point shown, one filled, one open
+                    {
+                        type: "line",
+                        kind: "line",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -5],
+                                color: "green",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -3],
+                                color: "green",
+                                filled: false,
+                            },
+                        ],
+                        color: "green",
+                        lineStyle: "solid",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: true,
+                    },
+                    // Dashed line, one point shown
+                    {
+                        type: "line",
+                        kind: "line",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -6],
+                                color: "blue",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -4],
+                                color: "purple",
+                                filled: false,
+                            },
+                        ],
+                        color: "purple",
+                        lineStyle: "dashed",
+                        showArrows: false,
+                        showStartPoint: true,
+                        showEndPoint: false,
+                    },
+                    // Show arrows
+                    {
+                        type: "line",
+                        kind: "line",
+                        points: [
+                            {
+                                type: "point",
+                                coord: [-7, -7],
+                                color: "blue",
+                                filled: true,
+                            },
+                            {
+                                type: "point",
+                                coord: [2, -5],
+                                color: "blue",
+                                filled: false,
+                            },
+                        ],
+                        color: "blue",
+                        lineStyle: "solid",
+                        showArrows: true,
+                        showStartPoint: false,
+                        showEndPoint: false,
+                    },
+                ],
+            },
+            type: "interactive-graph",
+            version: {
+                major: 0,
+                minor: 0,
+            },
+        },
+    },
+};
+
+// TODO(LEMS-1928): Add data for all locked rays, similar to the ones above.