+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/movable-point.tsx
@@ -1,9 +1,10 @@
 import {color as WBColor} from "@khanacademy/wonder-blocks-tokens";
+import Tooltip from "@khanacademy/wonder-blocks-tooltip";
 import {useMovable} from "mafs";
 import * as React from "react";
 import {useRef} from "react";
 
-import useGraphState from "../../reducer/use-graph-state";
+import useGraphConfig from "../../reducer/use-graph-config";
 import {snap} from "../../utils";
 import {useTransform} from "../use-transform";
 
@@ -20,27 +21,61 @@ type Props = {
 const hitboxSizePx = 48;
 
 export const StyledMovablePoint = (props: Props) => {
-    const {state} = useGraphState();
+    const {range, snapStep, markings, showTooltips} = useGraphConfig();
     const hitboxRef = useRef<SVGCircleElement>(null);
     const {point, onMove, color = WBColor.blue} = props;
+
+    // WB Tooltip requires a color name for the background color.
+    // Since the color in props is a hex value, a reverse lookup is needed.
+    const wbColorName = (Object.entries(WBColor).find(
+        ([_, value]) => value === color,
+    )?.[0] ?? "blue") as keyof typeof WBColor;
 
     const {dragging} = useMovable({
         gestureTarget: hitboxRef,
         point,
         onMove,
-        constrain: (p) => snap(state.snapStep, p),
+        constrain: (p) => snap(snapStep, p),
     });
     const pointClasses = `movable-point ${dragging ? "movable-point--dragging" : ""}`;
 
     const [[x, y]] = useTransform(point);
 
-    return (
+    const [xMin, xMax] = range[0];
+    const [yMin, yMax] = range[1];
+
+    const [[verticalStartX]] = useTransform([xMin, 0]);
+    const [[verticalEndX]] = useTransform([xMax, 0]);
+    const [[_, horizontalStartY]] = useTransform([0, yMin]);
+    const [[__, horizontalEndY]] = useTransform([0, yMax]);
+
+    const showHairlines = dragging && markings !== "none";
+    const hairlines = (
+        <g>
+            <line
+                x1={verticalStartX}
+                y1={y}
+                x2={verticalEndX}
+                y2={y}
+                stroke={color}
+            />
+            <line
+                x1={x}
+                y1={horizontalStartY}
+                x2={x}
+                y2={horizontalEndY}
+                stroke={color}
+            />
+        </g>
+    );
+
+    const svgForPoint = (
         <g
             ref={hitboxRef}
             className={pointClasses}
             tabIndex={0}
             style={{"--movable-point-color": color} as any}
-            data-test-id="movable-point"
+            data-testid="movable-point"
         >
             <circle
                 className="movable-point-hitbox"
@@ -56,8 +91,27 @@ export const StyledMovablePoint = (props: Props) => {
                 cx={x}
                 cy={y}
                 style={{fill: color}}
-                data-test-id="movable-point__center"
+                data-testid="movable-point__center"
             />
         </g>
+    );
+
+    return (
+        <>
+            {showHairlines && hairlines}
+
+            {showTooltips ? (
+                <Tooltip
+                    autoUpdate={true}
+                    backgroundColor={wbColorName}
+                    content={`(${point[0]}, ${point[1]})`}
+                    contentStyle={{color: "white"}}
+                >
+                    {svgForPoint}
+                </Tooltip>
+            ) : (
+                svgForPoint
+            )}
+        </>
     );
 };