+++ packages/perseus/src/widgets/interactive-graphs/grid.tsx
@@ -66,8 +66,20 @@ export const Grid = (props: GridProps) => {
                 xAxis={axisOptions(props, 0)}
                 yAxis={axisOptions(props, 1)}
             />
-            <AxisTicks range={props.range} tickStep={props.tickStep} />
             {props.markings === "graph" && <AxisArrows />}
+            {
+                // Only render the axis ticks and arrows if the markings are set to a full "graph"
+                props.markings === "graph" && (
+                    <>
+                        <AxisTicks
+                            range={props.range}
+                            tickStep={props.tickStep}
+                            gridStep={props.gridStep}
+                        />
+                        <AxisArrows />
+                    </>
+                )
+            }
         </>
     );
 };