+++ packages/perseus/src/widgets/interactive-graphs/mafs-graph.tsx
@@ -21,7 +21,7 @@ import {
     getGradableGraph,
     initializeGraphState,
 } from "./reducer/interactive-graph-state";
-import {GraphStateContext} from "./reducer/use-graph-state";
+import {GraphConfigContext} from "./reducer/use-graph-config";
 
 import type {InteractiveGraphState, InteractiveGraphProps} from "./types";
 import type {Widget} from "../../renderer";
@@ -41,6 +41,7 @@ export type Props = {
     containerSizeClass: InteractiveGraphProps["containerSizeClass"];
     markings: InteractiveGraphProps["markings"];
     onChange: InteractiveGraphProps["onChange"];
+    showTooltips: Required<InteractiveGraphProps["showTooltips"]>;
 };
 
 const renderGraph = (props: {
@@ -94,6 +95,7 @@ export const MafsGraph = (props: MafsGraphProps) => {
     const [width, height] = props.box;
 
     const prevState = useRef<InteractiveGraphState>(state);
+
     useEffect(() => {
         if (prevState.current !== state) {
             props.onChange({graph: state});
@@ -119,10 +121,12 @@ export const MafsGraph = (props: MafsGraphProps) => {
     }, [dispatch, xMinRange, xMaxRange, yMinRange, yMaxRange]);
 
     return (
-        <GraphStateContext.Provider
+        <GraphConfigContext.Provider
             value={{
-                state,
-                dispatch,
+                range: state.range,
+                snapStep: state.snapStep,
+                markings: props.markings,
+                showTooltips: !!props.showTooltips,
             }}
         >
             <View
@@ -184,6 +188,6 @@ export const MafsGraph = (props: MafsGraphProps) => {
                     </Mafs>
                 </View>
             </View>
-        </GraphStateContext.Provider>
+        </GraphConfigContext.Provider>
     );
 };