+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @khanacademy/perseus-editor
 
+## 5.3.0
+
+### Minor Changes
+
+-   [#1182](https://github.com/Khan/perseus/pull/1182) [`066719c82`](https://github.com/Khan/perseus/commit/066719c82cbb934287082961de93abb427c7600e) Thanks [@nishasy](https://github.com/nishasy)! - View locked lines on Interactive Graph
+
+*   [#1179](https://github.com/Khan/perseus/pull/1179) [`437bce7be`](https://github.com/Khan/perseus/commit/437bce7be32cbb03e0926719df0e03c193ba6df5) Thanks [@jeremywiebe](https://github.com/jeremywiebe)! - Change testId to render the default Testing Library HTML attribute: data-testid (was data-test-id)
+
+### Patch Changes
+
+-   [#1190](https://github.com/Khan/perseus/pull/1190) [`da6238a3a`](https://github.com/Khan/perseus/commit/da6238a3ad526c492476287c7bd7edd6ef04e4cb) Thanks [@handeyeco](https://github.com/handeyeco)! - Bugfix: remount graph in editor when a major change happens
+
+*   [#1177](https://github.com/Khan/perseus/pull/1177) [`f419d18fb`](https://github.com/Khan/perseus/commit/f419d18fbc76ff87720d9d49562c0b43ed1c2b33) Thanks [@nishasy](https://github.com/nishasy)! - LockedPointSettings update: Allow toggleable points, refactor so that it's easy to add lines later
+
+-   [#1194](https://github.com/Khan/perseus/pull/1194) [`09627c8dd`](https://github.com/Khan/perseus/commit/09627c8dde74136234a97a6e708e783797daf511) Thanks [@jeremywiebe](https://github.com/jeremywiebe)! - Enable M1 locked features in Perseus Editor storybook stories
+
+*   [#1178](https://github.com/Khan/perseus/pull/1178) [`ba0f36561`](https://github.com/Khan/perseus/commit/ba0f36561bf556b7bd0f85eb98be25c834643022) Thanks [@jeremywiebe](https://github.com/jeremywiebe)! - Move locked features for interactive-graph behind new flag named "interactive-graph-locked-features-m1"
+
+*   Updated dependencies [[`ba0f36561`](https://github.com/Khan/perseus/commit/ba0f36561bf556b7bd0f85eb98be25c834643022), [`a1e3fdfee`](https://github.com/Khan/perseus/commit/a1e3fdfee5768209a9d96749e78c5fd0e05f7dfb), [`066719c82`](https://github.com/Khan/perseus/commit/066719c82cbb934287082961de93abb427c7600e), [`437bce7be`](https://github.com/Khan/perseus/commit/437bce7be32cbb03e0926719df0e03c193ba6df5), [`f419d18fb`](https://github.com/Khan/perseus/commit/f419d18fbc76ff87720d9d49562c0b43ed1c2b33)]:
+    -   @khanacademy/perseus@21.6.0
+    -   @khanacademy/math-input@17.5.0
+
+## 5.2.14
+
+### Patch Changes
+
+-   Updated dependencies [[`661964d40`](https://github.com/Khan/perseus/commit/661964d40ba86e6b477ab87224692361f7aaaa27)]:
+    -   @khanacademy/perseus@21.5.0
+
+## 5.2.13
+
+### Patch Changes
+
+-   Updated dependencies [[`a3d724b86`](https://github.com/Khan/perseus/commit/a3d724b8657edd78266335e236f8eba5f5f1b9dd), [`dd97c422a`](https://github.com/Khan/perseus/commit/dd97c422a7d40a3786075d8c6988509fc2c43012), [`e3063ddc7`](https://github.com/Khan/perseus/commit/e3063ddc76ad22469ca5a8557bee78216869f5d3), [`3da51638b`](https://github.com/Khan/perseus/commit/3da51638b513e10c42da421667c3d877c5dcc161), [`0f80f4089`](https://github.com/Khan/perseus/commit/0f80f4089a6000301fc63ceef24646bd05a26db6), [`ccead133f`](https://github.com/Khan/perseus/commit/ccead133fa30e1d93e02fd9a7c9e544c750e55ab)]:
+    -   @khanacademy/perseus@21.4.0
+
+## 5.2.12
+
+### Patch Changes
+
+-   [#1174](https://github.com/Khan/perseus/pull/1174) [`108cb3bc4`](https://github.com/Khan/perseus/commit/108cb3bc4b9c4d734c4f8cd424e55f8c14c77009) Thanks [@handeyeco](https://github.com/handeyeco)! - Bugfix: allow selecting financial calculator options again
+
+## 5.2.11
+
+### Patch Changes
+
+-   Updated dependencies [[`53fa89b50`](https://github.com/Khan/perseus/commit/53fa89b50240309dcf56919b77715e135682e0ee), [`9c5363795`](https://github.com/Khan/perseus/commit/9c53637953ff0267306bdc98175be9770bf2f098), [`ce547c5ec`](https://github.com/Khan/perseus/commit/ce547c5ec1ef0d17e0adebe24532c93007e82c4d)]:
+    -   @khanacademy/perseus@21.3.1
+
 ## 5.2.10
 
 ### Patch Changes